fix: clippy warnings for strict CI gate

+++ src-tauri/src/history/claude.rs
@@ -398,7 +398,7 @@ pub(super) fn parse_iso_ms(ts: &str) -> Option<u128> {
 fn days_from_civil(y: i64, m: u32, d: u32) -> i64 {
     let y = if m <= 2 { y - 1 } else { y };
     let era = if y >= 0 { y } else { y - 399 } / 400;
-    let yoe = (y - era * 400) as i64;
+    let yoe = y - era * 400;
     let m = m as i64;
     let d = d as i64;
     let doy = (153 * (if m > 2 { m - 3 } else { m + 9 }) + 2) / 5 + d - 1;
